+++ src/com/shopper2/vista/EditarPedido.java
@@ -4,83 +4,40 @@
  */
 package com.shopper2.vista;
 
-import com.shopper2.modelo.dao.PedidoDao;
-import com.shopper2.modelo.dao.ProductoDao;
-import com.shopper2.modelo.dao.RepartidorDao;
-import com.shopper2.modelo.pedido.Pedido;
-import com.shopper2.modelo.productos.IAgregarProducto;
-import com.shopper2.modelo.productos.IProducto;
-import com.shopper2.modelo.productos.Producto;
-import com.shopper2.modelo.repartidores.Repartidor;
+import com.shopper2.controlador.EditarPedidoControlador;
 
 import javax.swing.*;
-import javax.swing.table.DefaultTableModel;
-import java.sql.Date;
-import java.util.ArrayList;
-import java.util.Map;
 
-public class EditarPedido extends javax.swing.JFrame implements IAgregarProducto {
+public class EditarPedido extends javax.swing.JFrame {
+    EditarPedidoControlador controlador;
 
-    private final int codpe;
 
     /**
      * Creates new form NuevoPedido
      */
-    public EditarPedido(int codpe) {
-        this.codpe = codpe;
+    public EditarPedido(EditarPedidoControlador controlador) {
+        this.controlador = controlador;
         initComponents();
-        initRepartidores();
-        initPedido();
     }
 
-    private void initPedido() {
-        Pedido pedido = PedidoDao.getInstance().buscar(codpe);
-        tNombreCliente.setText(pedido.getNomCliente());
-        tDireccionCliente.setText(pedido.getDireccionCliente());
-        tFechaPedido.setText(String.valueOf(pedido.getFecha()));
-        int elementosBox = boxRepartidores.getItemCount();
-        for (int i = 0; i < elementosBox; i++) {
-            String elemento = boxRepartidores.getItemAt(i);
-            String[] parts = elemento.split(" - ");
-            int codr = Integer.parseInt(parts[0]);
-            if (codr == pedido.getRepartidor().getCodr()) {
-                boxRepartidores.setSelectedIndex(i);
-                break;
-            }
-        }
-
-        for (Map.Entry<IProducto, Integer> entrada : pedido.getProductos().entrySet()) {
-            DefaultTableModel model = (DefaultTableModel) tablaProductos.getModel();
-            Object[] datos = new Object[4];
-            datos[0] = entrada.getKey().getCodpr();
-            datos[1] = entrada.getKey().getNombreProducto();
-            datos[2] = entrada.getKey().getCategoria();
-            datos[3] = entrada.getValue();
-            model.addRow(datos);
-        }
+    public JComboBox<String> getBoxRepartidores() {
+        return boxRepartidores;
     }
 
-    private void initRepartidores() {
-        DefaultComboBoxModel<String> model = new DefaultComboBoxModel<>();
-        ArrayList<Repartidor> repartidores = RepartidorDao.getInstance().obtenerRepartidores();
-        for (Repartidor repartidor : repartidores) {
-            model.addElement(repartidor.getCodr() + " - " + repartidor.getNomr());
-        }
-        boxRepartidores.setModel(model);
+    public JTextField gettDireccionCliente() {
+        return tDireccionCliente;
     }
 
-    public void agregarProducto(Producto producto, int cantidad) {
+    public JTextField gettFechaPedido() {
+        return tFechaPedido;
+    }
 
-        if (producto != null) {
-            DefaultTableModel model = (DefaultTableModel) tablaProductos.getModel();
-            Object[] datos = new Object[4];
-            datos[0] = producto.getCodpr();
-            datos[1] = producto.getNombreProducto();
-            datos[2] = producto.getCategoria();
-            datos[3] = cantidad;
-            model.addRow(datos);
+    public JTextField gettNombreCliente() {
+        return tNombreCliente;
+    }
 
-        }
+    public JTable getTablaProductos() {
+        return tablaProductos;
     }
 
     /**
@@ -264,72 +221,19 @@ public class EditarPedido extends javax.swing.JFrame implements IAgregarProducto
     }// </editor-fold>//GEN-END:initComponents
 
     private void bAddProductoActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_bAddProductoActionPerformed
-        AgregarProductos agregarProductos = new AgregarProductos(this);
-        agregarProductos.setVisible(true);
+        controlador.abrirAgregarProductos();
     }//GEN-LAST:event_bAddProductoActionPerformed
 
     private void bDelProductoActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_bDelProductoActionPerformed
-        DefaultTableModel model = (DefaultTableModel) tablaProductos.getModel();
-        int filaSeleccionada = tablaProductos.getSelectedRow();
-
-        if (filaSeleccionada != -1) {
-            // Obtener el valor de la columna 0 (suponiendo que sea la columna primaria)
-            int productoId = (int) model.getValueAt(filaSeleccionada, 0);
-
-            // Eliminar la fila seleccionada del modelo de tabla
-            model.removeRow(filaSeleccionada);
-
-            // Eliminar el producto de la base de datos utilizando el ID
-            ProductoDao.getInstance().eliminar(productoId);
-        }
+        controlador.borrarProducto(tablaProductos);
     }//GEN-LAST:event_bDelProductoActionPerformed
 
     private void bCancelarActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_bCancelarActionPerformed
-        //ListaPedidos listaPedidos = new ListaPedidos();
-        //listaPedidos.setVisible(true);
-        dispose();
+        controlador.abrirListaPedidos();
     }//GEN-LAST:event_bCancelarActionPerformed
 
     private void bModificarActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_bModificarActionPerformed
-        //Objeto de tipo Pedido para recoger la información introducida por el usuario
-        Pedido pedido = new Pedido();
-        pedido.setCodpe(codpe);
-        //Almacenamiento del dato nombre del cliente
-        pedido.setNomCliente(tNombreCliente.getText());
-        //Almacenamiento del dato dirección del cliente
-        pedido.setDireccionCliente(tDireccionCliente.getText());
-        //Almacenamiento del dato fecha del pedido
-        pedido.setFecha(Date.valueOf(tFechaPedido.getText()));
-        //Almacenamiento del dato repartidor
-        String selectedItem = (String) boxRepartidores.getSelectedItem();
-        String[] parts = selectedItem.split(" - ");
-        int codr = Integer.parseInt(parts[0]);
-        Repartidor repartidor = RepartidorDao.getInstance().buscar(codr);
-        pedido.setRepartidor(repartidor);
-        //Se obtiene el modelo de la tabla asociada a la tablaProductos
-        DefaultTableModel model = (DefaultTableModel) tablaProductos.getModel();
-        //Se obtiene la cantidad de filas que tiene la tabla
-        int cantidadDeFilas = model.getRowCount();
-        //Se inicia el bucle que recorre cada fila de la tablaProductos
-        for (int fila = 0; fila < cantidadDeFilas; fila++) {
-            //Se extrae el valor de la columna 0 de cada fila y se guarda en la variable codpr
-            int codpr = (Integer) (model.getValueAt(fila, 0));
-            //Se extrae el valor de la columna 3 de cada fila y se guarda en la variable cantidad
-            int cantidad = (Integer) (model.getValueAt(fila, 3));
-            //se crea un nuevo objeto Producto y se establece su código (codpr).
-            //Finalmente, se agrega el producto y la cantidad al objeto Pedido
-            //utilizando el método addProducto().
-            Producto producto = new Producto();
-            producto.setCodpr(codpr);
-            pedido.addProducto(producto, cantidad);
-        }
-
-
-        if (PedidoDao.getInstance().modificar(pedido)) {
-            //ListaPedidos pedidos = new ListaPedidos();
-            //pedidos.setVisible(true);
-            dispose();
-        }
+        controlador.modificarPedido();
     }//GEN-LAST:event_bModificarActionPerformed
 
     // Variables declaration - do not modify//GEN-BEGIN:variables
